+++ src/edu/uclouvain/core/nodus/compute/assign/ExactMFAssignment.java
@@ -76,6 +76,9 @@ public class ExactMFAssignment extends Assignment {
       assignmentParameters.setDurationFunctions(true);
     }
 
+    // Get the max detour reference mode
+    getMaxDetourReferenceMode();
+
     // Generate a virtual network
     virtualNet = new VirtualNetwork(assignmentParameters);