+++ src/edu/uclouvain/core/nodus/database/dbf/ImportDBF.java
@@ -115,7 +115,6 @@ public class ImportDBF {
     try {
       PreparedStatement prepStmt = jdbcConnection.prepareStatement(sqlStmt);
 
-      boolean oldAutoCommit = jdbcConnection.getAutoCommit();
       int batchSize = 0;
 
       while (dbfReader.hasNextRecord()) {